[PERFORMANCE] 리액트 성능 최적화 (#219)
* perf: Waiting page에서 WaitingListContainer와 WaitingBottomBarContainer 컴포넌트 분리

- 다른 사용자 입장, 퇴장 시 불필요한 헤더의 리렌더링 없앰

* perf: ProfileCardContainer 컴포넌트 분리 및 memo로 메모이제이션 수행

* perf: 투표 이벤트 최적화

* perf: 게임 페이지의 모달 분리

- 모달 상태 값들 모달 컴포넌트에서 관리
- 메모이제이션 수행

* perf: 투표 및 능력 사용 최적화 (ing)

투표와 능력 관련 데이터들을 Game 페이지 하위 컴포넌트들이 관리하도록 수정
- useGame 커스텀 훅에서 players 데이터 삭제 및 관련 데이터 useAbility와 useJobMemo로 이동
- players 데이터 삭제에 맞게 관련 내용들 수정

* perf: 투표 및 능력 사용 최적화 (ing)

- Game 페이지에서 players 데이터 삭제
- useAbility 커스텀 훅 하위 컴포넌트 (LeftSideContainer)로 이동

* refactor: 내가 죽었는지를 나타내는 데이터 생성

- `useRef` API 사용해서 내가 죽었는지를 나타내는 데이터 생성
- 각 모달과 useAbilityButtonList 로 전달
