+++ packages/client/src/containers/LeftSideContainer.tsx
@@ -1,66 +1,59 @@
-import { FC, useEffect } from 'react';
+import { FC, useEffect, useRef } from 'react';
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import { useHistory } from 'react-router-dom';
 
+import * as EVENT from '@mafia/domain/constants/event';
 import useModal from '@hooks/useModal';
 import { titleActive, white, grey1 } from '@constants/index';
 import { RoomOutIcon, AudioOffIcon, AudioOnIcon } from '@components/Icon';
-import ConfirmModal from '@components/Modal/ConfirmModal';
 import { IconButton, ButtonSizeList, ButtonThemeList } from '@components/Button';
 import { GAME_DAY_MP3 } from '@constants/audio';
 import useAudio from '@src/hooks/useAudio';
-import { Player, Selected } from '@src/types';
-import useExecutionModal from '@src/hooks/useExecutionModal';
+import { User } from '@mafia/domain/types/user';
 import AbilityButtonList from '@src/lists/AbilityButtonList';
 import { useUserInfo } from '@src/contexts/userInfo';
 import Timer from '@src/components/Timer';
+import useAbility from '@hooks/useAbility';
+import { Event } from '@src/types';
+import useSocketEvent from '@hooks/useSocketEvent';
+import { useSocketContext } from '@src/contexts/socket';
+import GamePageModalContainer from './GamePageModalContainer';
 
 type PropType = {
-  players: Player[];
-  mafias: string[];
-  selected: Selected;
+  initPlayers: User[];
   isNight: boolean;
-  getSelectedImg: any;
-  emitAbility: any;
   roomName: string;
+  myJob: string;
 };
 
-const LeftSideContainer: FC<PropType> = ({
-  players,
-  mafias,
-  selected,
-  isNight,
-  getSelectedImg,
-  emitAbility,
-  roomName,
-}) => {
-  const history = useHistory();
-
+const LeftSideContainer: FC<PropType> = ({ initPlayers, isNight, myJob, roomName }) => {
+  const { socketRef } = useSocketContext();
   const { userInfo } = useUserInfo();
+  const amIDead = useRef<boolean>(false);
 
+  const setMyDeadState = (playerName: string) => {
+    if (userInfo?.userName === playerName) {
+      amIDead.current = true;
+    }
+  };
+
+  const executionEvent: Event = { event: EVENT.EXECUTION, handler: setMyDeadState };
+  const killEvent: Event = { event: EVENT.PUBLISH_VICTIM, handler: setMyDeadState };
+  const exitEvent: Event = { event: EVENT.EXIT, handler: setMyDeadState };
+  useSocketEvent(socketRef, [executionEvent, killEvent, exitEvent], [amIDead.current]);
+
+  const { players, mafias, selected, emitAbility, getSelectedImg } = useAbility(
+    initPlayers,
+    isNight,
+    myJob,
+  );
   const {
     isModalOpen: isRoomOutModalOpen,
     openModal: openRoomOutModal,
     closeModal: closeRoomOutModal,
   } = useModal();
 
-  const {
-    isModalOpen: isExecutionModalOpen,
-    maxVotedPlayer,
-    closeModal: closeExecutionModal,
-    executionHandler,
-  } = useExecutionModal();
-
   const { playing, updateLoop, toggle, pause } = useAudio(GAME_DAY_MP3);
-
-  const roomOutHandler = () => {
-    history.push('/rooms');
-    closeRoomOutModal();
-  };
-
-  const amIDead = () =>
-    players.find(({ userName: playerName }) => playerName === userInfo?.userName)?.isDead;
 
   useEffect(() => {
     updateLoop(true);
@@ -73,21 +66,11 @@ const LeftSideContainer: FC<PropType> = ({
 
   return (
     <div css={leftSideContainerStyle}>
-      <ConfirmModal
-        isOpen={isRoomOutModalOpen}
-        onRequestClose={closeRoomOutModal}
-        eventHandler={roomOutHandler}
-        closeModal={closeRoomOutModal}
-      >
-        <p>진행중인 게임을 포기하고 나가시겠습니까?</p>
-      </ConfirmModal>
-      <ConfirmModal
-        isOpen={amIDead() ? false : isExecutionModalOpen}
-        eventHandler={executionHandler}
-        closeModal={closeExecutionModal}
-      >
-        <p>{maxVotedPlayer}을(를) 투표로 처형할까요?</p>
-      </ConfirmModal>
+      <GamePageModalContainer
+        amIDead={amIDead}
+        isRoomOutModalOpen={isRoomOutModalOpen}
+        closeRoomOutModal={closeRoomOutModal}
+      />
       <div css={Style}>
         <img
           src={isNight ? '/assets/images/moon.png' : '/assets/images/sun.png'}
@@ -119,7 +102,7 @@ const LeftSideContainer: FC<PropType> = ({
         selected={selected}
         emitAbility={emitAbility}
         getSelectedImg={getSelectedImg}
-        amIDead={amIDead() ?? false}
+        amIDead={amIDead ?? false}
       />
     </div>
   );
