+++ packages/client/src/containers/RightSideContainer.tsx
@@ -2,7 +2,6 @@ import { FC, useState } from 'react';
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 
-import { Player } from '@src/types';
 import { grey1, titleActive, white, JOB_DICT } from '@constants/index';
 import { Modal } from '@components/Modal';
 import { SearchIcon } from '@components/Icon';
@@ -12,13 +11,12 @@ import MemoList from '@src/lists/MemoList';
 import { User } from '@mafia/domain/types/user';
 
 type PropType = {
-  players: Player[];
   isNight: boolean;
   myJob: string;
   initPlayers: User[];
 };
 
-const RightSideContainer: FC<PropType> = ({ initPlayers, players, isNight, myJob }) => {
+const RightSideContainer: FC<PropType> = ({ initPlayers, isNight, myJob }) => {
   const [showModal, setShowModal] = useState(false);
 
   const handleClick = () => {
@@ -35,7 +33,7 @@ const RightSideContainer: FC<PropType> = ({ initPlayers, players, isNight, myJob
         </div>
       </div>
       <hr css={hrStyle} />
-      <MemoList players={players} isNight={isNight} initPlayers={initPlayers} />
+      <MemoList isNight={isNight} initPlayers={initPlayers} />
       <hr css={hrStyle} />
       {showModal ? (
         <div css={modalWrapperStyle}>
