+++ packages/client/src/lists/WaitingList.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import { ProfileCard } from '@src/components/Card';
 import { PlayerInfo } from '@mafia/domain/types/user';
+import ProfileCardContainer from '@src/containers/ProfileCardContainer';
 
 interface Prop {
   userList: PlayerInfo[];
@@ -11,15 +11,13 @@ interface Prop {
 const WaitingList: FC<Prop> = ({ userList }) => (
   <div css={containerStyle}>
     {userList.map(({ userName, profileImg, isHost, isReady }) => (
-      <div key={userName} css={wrapperStyle}>
-        <img css={decoStyle(isHost)} src="\assets\icons\host.png" alt="host" />
-        <ProfileCard
-          userName={userName}
-          status={isHost ? 'WAITING' : 'READY'}
-          profileImg={profileImg}
-          fill={isReady}
-        />
-      </div>
+      <ProfileCardContainer
+        key={userName}
+        userName={userName}
+        profileImg={profileImg}
+        isHost={isHost}
+        isReady={isReady}
+      />
     ))}
   </div>
 );
@@ -39,21 +37,6 @@ const containerStyle = css`
   &::-webkit-scrollbar {
     display: none;
   }
-`;
-
-const wrapperStyle = css`
-  position: relative;
-`;
-
-const decoStyle = (isHost: boolean) => (isHost ? iconStyle : iconNone);
-
-const iconStyle = css`
-  position: absolute;
-  right: -25px;
-  top: -35px;
-`;
-const iconNone = css`
-  display: none;
 `;
 
 export default WaitingList;
