+++ packages/socket/package.json
@@ -18,7 +18,6 @@
     "@types/jsonwebtoken": "^8.5.6",
     "@typescript-eslint/eslint-plugin": "^5.2.0",
     "@typescript-eslint/parser": "^5.2.0",
-    "artillery": "^2.0.0-5",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
